spring-boot-crud-operation/src/main/java/com/javatpoint/tutorialsPoint/MathApplication.java
package com.javatpoint.tutorialsPoint;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class MathApplication {
	
	@Autowired
	private CalculatorService calcService;

	public double add(double input1, double input2) {
		return calcService.add(input1, input2);
	}

	public double subtract(double input1, double input2) {
		return calcService.subtract(input1, input2);
	}

	public double multiply(double input1, double input2) {
		return calcService.multiply(input1, input2);
	}

	public double divide(double input1, double input2) {
		return calcService.divide(input1, input2);
	}
}
